Split similarity prep progress refresh
